+++ src/app/(main)/(mobile)/me/features/Cate.tsx
@@ -1,6 +1,5 @@
 'use client';
 
-import { useTheme } from 'antd-style';
 import { useRouter } from 'next/navigation';
 import { memo } from 'react';
 import { Flexbox } from 'react-layout-kit';
@@ -11,12 +10,11 @@ import Cell from '@/components/Cell';
 import Divider from '@/components/Cell/Divider';
 
 const SettingCate = memo(() => {
-  const theme = useTheme();
   const settingItems = useCategory({ mobile: true });
   const router = useRouter();
 
   return (
-    <Flexbox style={{ background: theme.colorBgContainer }} width={'100%'}>
+    <Flexbox width={'100%'}>
       {settingItems?.map(({ key, icon, label, type }: any, index) => {
         if (type === 'divider') return <Divider key={index} />;
         return (