+++ src/store/user/store.ts
@@ -6,19 +6,28 @@ import { StateCreator } from 'zustand/vanilla';
 import { isDev } from '@/utils/env';
 
 import { type UserState, initialState } from './initialState';
+import { type UserAuthAction, createAuthSlice } from './slices/auth/action';
 import { type CommonAction, createCommonSlice } from './slices/common/action';
 import { type PreferenceAction, createPreferenceSlice } from './slices/preference/action';
 import { type SettingsAction, createSettingsSlice } from './slices/settings/actions';
+import { type SyncAction, createSyncSlice } from './slices/sync/action';
 
 //  ===============  聚合 createStoreFn ============ //
 
-export type UserStore = CommonAction & UserState & SettingsAction & PreferenceAction;
+export type UserStore = SyncAction &
+  UserState &
+  SettingsAction &
+  PreferenceAction &
+  UserAuthAction &
+  CommonAction;
 
 const createStore: StateCreator<UserStore, [['zustand/devtools', never]]> = (...parameters) => ({
   ...initialState,
-  ...createCommonSlice(...parameters),
+  ...createSyncSlice(...parameters),
   ...createSettingsSlice(...parameters),
   ...createPreferenceSlice(...parameters),
+  ...createAuthSlice(...parameters),
+  ...createCommonSlice(...parameters),
 });
 
 //  ===============  实装 useStore ============ //